+++ ctrlf.el
@@ -1181,6 +1181,7 @@ display an error message and do not search."
             (occur regexp)))
       (ctrlf--message "No input available"))))
 
+;;;###autoload
 (defun ctrlf-forward-fuzzy ()
   "Fuzzy search forward for literal string.
 If already in a search, go to next candidate, or if no input then
@@ -1189,6 +1190,7 @@ change back to fuzzy search."
   (interactive)
   (ctrlf-forward 'fuzzy))
 
+;;;###autoload
 (defun ctrlf-backward-fuzzy ()
   "Fuzzy search backward for literal string.
 If already in a search, go to previous candidate, or if no input
@@ -1197,6 +1199,7 @@ change back to fuzzy search."
   (interactive)
   (ctrlf-backward 'fuzzy))
 
+;;;###autoload
 (defun ctrlf-forward-fuzzy-regexp ()
   "Fuzzy search forward for literal string.
 If already in a search, go to next candidate, or if no input then
@@ -1205,6 +1208,7 @@ search, change back to fuzzy-regexp search."
   (interactive)
   (ctrlf-forward 'fuzzy-regexp))
 
+;;;###autoload
 (defun ctrlf-backward-fuzzy-regexp ()
   "Fuzzy search backward for literal string.
 If already in a search, go to previous candidate, or if no input