+++ CHANGELOG.md
@@ -9,12 +9,17 @@ The format is based on [Keep a Changelog].
   minibuffer to search the current user input and the minibuffer
   history. Note that this functionality is not compatible with
   Selectrum. See [#41].
+* We now disable `blink-matching-paren` automatically when entering
+  CTRLF, because (for some reason) it uses `sit-for` to display its
+  blink and this of course causes search results to be delayed
+  ([#45]).
 
 ### Bugs fixed
 * CTRLF previously caused an error during Emacs startup unless the
   `map` library was already loaded. This has been fixed ([#49]).
 
 [#41]: https://github.com/raxod502/ctrlf/issues/41
+[#45]: https://github.com/raxod502/ctrlf/issues/45
 [#49]: https://github.com/raxod502/ctrlf/issues/49
 
 ## 1.0 (released 2020-03-31)